+++ src/classes/tmpltown.cpp
@@ -33,14 +33,14 @@ unsigned char& cBigTemplTown::lighting(size_t i, size_t r){
 	return _lighting[i][r];
 }
 
-short cBigTemplTown::max_dim() const {
+size_t cBigTemplTown::max_dim() const {
 	return 0; // not sure what they are yet.
 }
 
-short cBigTemplTown::max_monst() const {
+size_t cBigTemplTown::max_monst() const {
 	return 30;
 }
 
-short cBigTemplTown::max_items() const {
+size_t cBigTemplTown::max_items() const {
 	return 64;
 }
