+++ src/classes/tmpltown.h
@@ -45,9 +45,9 @@ public:
 	ter_num_t& terrain(size_t x, size_t y);
 	cCreature& creatures(size_t i);
 	unsigned char& lighting(size_t i, size_t r);
-	short max_dim() const;
-	short max_monst() const;
-	short max_items() const;
+	size_t max_dim() const;
+	size_t max_monst() const;
+	size_t max_items() const;
 	void writeTerrainTo(std::ostream& file);
 	void readTerrainFrom(std::istream& file);
 	explicit cBigTemplTown(cScenario& scenario, bool init_strings = false);
@@ -62,9 +62,9 @@ public:
 	ter_num_t& terrain(size_t x, size_t y);
 	cCreature& creatures(size_t i);
 	unsigned char& lighting(size_t i, size_t r);
-	short max_dim() const;
-	short max_monst() const;
-	short max_items() const;
+	size_t max_dim() const;
+	size_t max_monst() const;
+	size_t max_items() const;
 	void writeTerrainTo(std::ostream& file);
 	void readTerrainFrom(std::istream& file);
 	explicit cMedTemplTown(cScenario& scenario, bool init_strings = false);
@@ -79,9 +79,9 @@ public:
 	ter_num_t& terrain(size_t x, size_t y);
 	cCreature& creatures(size_t i);
 	unsigned char& lighting(size_t i, size_t r);
-	short max_dim() const;
-	short max_monst() const;
-	short max_items() const;
+	size_t max_dim() const;
+	size_t max_monst() const;
+	size_t max_items() const;
 	void writeTerrainTo(std::ostream& file);
 	void readTerrainFrom(std::istream& file);
 	explicit cTinyTemplTown(cScenario& scenario, bool init_strings = false);
